Simplify Sveltekit Example
